Obstacle Avoiding Car
This is a program created for an Arduino obstacle-avoiding car. Using an ultrasonic sensor, the car stops after detecting any solid object less than 30 cm ahead of it. The car then reverses, turns left and right to check which direction is clear of obstacles, and continues moving in that direction.